LinkLocalAddressing: Fixing overriden LinkLocalAddressing Value.
Issue: This issue exist with set-ipv4 and set-ipv6 command.
While setting LinkLocalAddressing for both IPv4 and IPv6
one after another the LinkLocalAddressing only preserving
the last command value in this case it would be LinkLocalAddressing=ipv6
instead it should be LinkLocalAddressing=yes.

Fix: Addressing this issue in both commands now nmctl check for existing value
and based on that it will update the LinkLocalAddressing parameter.